requirements.txt
apiclient==1.0.3
certifi==2017.11.5
chardet==3.0.4
Django==2.0.1
django-reset-migrations==0.3.1
djangorestframework==3.7.7
enum34==1.1.6
futures==3.1.1
google-api-python-client==1.6.4
goslate==1.5.1
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
mysqlclient==1.3.12
Naked==0.1.31
numpy==1.14.1
oauth2==1.9.0.post1
oauth2client==4.1.2
oauthlib==2.0.6
pandas==0.22.0
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
python-dateutil==2.6.1
python-http-client==3.0.0
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
sendgrid==5.3.0
setproctitle==1.1.10
shellescape==3.4.1
simplejson==3.13.2
six==1.11.0
tweepy==3.5.0
uritemplate==3.0.0
urllib3==1.22
